ECS - Refactored MovementSystem to a physics-based model and updated related components.
- Updated `MovementSystem` to use acceleration, mass, and friction instead of direct velocity setting.
- Implemented a "soft speed limit" using aggressive drag for entities exceeding their maximum velocity.
- Modified `MovementSystem` to exclude entities with `NetworkSyncComponent`.
- Changed data types from `double` to `float` in `AccelerationComponent` and `MassComponent`.
- Expanded `MovementSystemTest` to cover acceleration, friction, soft speed limits, and remote entity exclusion.

+++ core/src/main/java/com/fatpiggies/game/model/ecs/components/AccelerationComponent.java
@@ -3,8 +3,8 @@ package com.fatpiggies.game.model.ecs.components;
 import com.badlogic.ashley.core.Component;
 
 public class AccelerationComponent implements Component {
-    public double baseMaxAcceleration;
-    public double currentMaxAcceleration;
-    public double ax;
-    public double ay;
+    public float baseMaxAcceleration;
+    public float currentMaxAcceleration;
+    public float ax;
+    public float ay;
 }
