+++ core/src/main/java/com/fatpiggies/game/model/ecs/components/MassComponent.java
@@ -3,6 +3,6 @@ package com.fatpiggies.game.model.ecs.components;
 import com.badlogic.ashley.core.Component;
 
 public class MassComponent implements Component {
-    public double baseMass;
-    public double currentMass;
+    public float baseMass;
+    public float currentMass;
 }
